+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
   - 🐞 [@mixte/components] 修复组件 `GridTable` 的表头分割线在某些情况下不显示的问题
   - 🐞 [@mixte/components] 修复组件 `InfiniteScroll` 的 `reset()` 方法重置后每页条数不一致的问题
+  - 🐞 [@mixte/snippets] 修复 `toggleThemeViewTransition` 动画取消后临时样式未及时清理的问题
 
 ## [v3.6.0-beta.5]
   - 📅 2026-04-16